My thoughts
Monero is open source

Monero is a grassroots community attracting the world's best cryptocurrency researchers, engineering, and creative talent. Over 240 developers have contributed to The Monero Project, including 30 core developers. As a project, Monero was launched without traditional venture capital funding, an Initial Coin Offering, or a premine. Today, Monero continues to thrive organically both as a project and a cryptocurrency without any institutional backing or corporate infrastructure. The Monero Project operates as a collective and it is not dependent on any one individual.